requirements.txt
certifi >= 14.05.14
six >= 1.10
python-dotenv == 0.20.0
python_dateutil >= 2.5.3
setuptools >= 21.0.0
urllib3 >= 1.15.1
